Hello world
The idea being explored is that we can load a resources json
from a ghost blog, but I'm not sure cross site scripting will
allow it to happen properly.  The file has to existbefore we
can know, and I don't want to spend time making a real one if
it's not going to work.